Update decidim to 0.9.2 (#199)
* Update decidim

* Use 0.9 in circle

* Use the test image

* Use the right version

* Add missing migrations

* Fix specs

* Don't test debates

* Extract results import call to a background job (#200)

* Extract results import call to a background job

* Fix results weight type

* Add import link

* Update to the latest commit

* Update to 0.9 stable

* Serialize the whole CSV content

* Fix specs

* Fix locales reference (#201)

* Fix locales path

* Force CSV content encoding

* Update to decidim 0.9.2

* Don't block when updating nicknames

+++ config/initializers/accountability_patches.rb
@@ -43,7 +43,7 @@ end
 
 Decidim::Accountability::Admin::ResultForm.class_eval do
   attribute :external_id, String
-  attribute :weight, Integer
+  attribute :weight, Float
 
   _validators[:description]
     .find { |v| v.is_a? TranslatablePresenceValidator }
